+++ src/functions/arithmetic/multiply.js
@@ -18,6 +18,8 @@ class LimeFunctionMultiply extends LimeFunction {
 
       'cl(var->expr)',
       'cr(var->expr)',
+      'cl(arg{expr}->expr)',
+      'cr(arg{expr}->expr)',
       'cl(bool->int)',
       'cr(bool->int)',
       'cb(int->rat,rat)',