+++ src/functions/bitwise/bitwiseNot.js
@@ -17,6 +17,7 @@ class LimeFunctionBitwiseNot extends LimeFunction {
       'tr(_-)',
 
       'cr(var->expr)',
+      'cr(arg{expr}->expr)',
       'cr(bool->int)',
 
       'er(int)',