+++ src/functions/comparison/greater.js
@@ -20,6 +20,8 @@ class LimeFunctionGreater extends LimeFunction {
 
       'cl(var->expr)',
       'cr(var->expr)',
+      'cl(arg{expr}->expr)',
+      'cr(arg{expr}->expr)',
       'cl(bool->int)',
       'cr(bool->int)',
       'cb(int->rat,rat)',