+++ src/functions/comparison/notEqual.js
@@ -19,6 +19,8 @@ class LimeFunctionNotEqual extends LimeFunction {
 
       'cl(var->expr)',
       'cr(var->expr)',
+      'cl(arg{expr}->expr)',
+      'cr(arg{expr}->expr)',
       'cl(bool->int)',
       'cr(bool->int)',
       'cb(int->rat,rat)',