+++ src/structs/expression.js
@@ -11,6 +11,8 @@ class LimeExpression {
     this.lime = lime;
 
     // Properties
+    this.name = args.hasOwnProperty('name') ? args.name : this.constructor.name.slice(12);
+    this.shorthand = args.hasOwnProperty('shorthand') ? args.shorthand : '';
     this.isSimple = args.hasOwnProperty('isSimple') && Types.isBoolean(args.isSimple) ? args.isSimple : true;
   }
 }