+++ docs/CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
-### v0.1.2
+### v0.2.0 (dev)
+- Added expressions: `Argument`
+- Added functions:
+  - Defining: `DefineBoolean`, `DefineInteger`, `DefineRational`
+  - Structural: `ColumnSplit`, `Omitted`
+- Modified existing functions to support new expression types
+### v0.1.2 (Aug 16, 2021)
 - Set up prompt and evaluate modes for input
 - Added commands: `Clear`, `Config`, `List`
 - Modified existing function to support boolean and non-boolean conversions
 - Set up crosstesting system and standardized process-based tests
-### v0.1.1
+### v0.1.1 (Aug 10, 2021)
 - Set up command system in prompt
 - Added commands: `About`, `Help`
 - Added expressions: `Boolean`, `Variable`
@@ -15,11 +21,12 @@
   - Comparison: `Equal`, `Greater`, `GreaterEqual`, `NotEqual`, `Smaller`, `SmallerEqual`
   - Counting: `Factorial`
   - Logical: `LogicalAnd`, `LogicalNot`, `LogicalOr`
+  - Structural: `Assign`
 - Modified existing functions to support new expression types
 - Improved engine independence of expression and function tests
 - Added tests: `Commands`
 - Removed tests: `Banner`
-### v0.1.0
+### v0.1.0 (Aug 6, 2021)
 - Set up engine framework, bundling system using Rollup and testing system using Jest
 - Set up prompt and web examples
 - Added expressions: `Integer`, `Rational`